VIRO-739: Adjusted text point to scale param, changed method signatures from float to int.
- Changed method sig. in VROText from float to int to avoid loss
of precision when calculating bounds for multiple lines for heights less than 1.

- Adjusted kTextPointToWorldScale to smaller value to reduce text size, text was too large.

+++ ViroRenderer/VROKnuthPlassFormatter.cpp
@@ -15,7 +15,7 @@
 int kInfinity = 10000;
 
 VROKnuthPlassFormatter::VROKnuthPlassFormatter(std::vector<std::shared_ptr<KPNode>> &nodes,
-                                               std::vector<int> &lineLengths,
+                                               std::vector<float> &lineLengths,
                                                float tolerance) :
     _nodes(nodes),
     _lineLengths(lineLengths) {
